// features/sdnr/odlux/odlux/apps/unmFaultManagementApp/src/models/unmFault.ts
/**
 * ============LICENSE_START========================================================================
 * ONAP : ccsdk feature sdnr wt odlux
 * =================================================================================================
 * Copyright (C) 2019 highstreet technologies GmbH Intellectual Property. All rights reserved.
 * =================================================================================================
 * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
 * in compliance with the License. You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software distributed under the License
 * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
 * or implied. See the License for the specific language governing permissions and limitations under
 * the License.
 * ============LICENSE_END==========================================================================
 */
export type FaultLog = {
  id?: string;
  nodeId: string;
  counter: number;
  objectId: string;
  timestamp: { value: string };
  problem: string;
  severity: null | 'Warning' | 'Minor' | 'Major' | 'Critical' | 'Indeterminate';
  sourceType?: string;
  comment?: string;
};


/**
 * Fault status count return 
 */
export type FaultsCountReturnType = {
  criticals: number;
  majors: number;
  minors: number;
  warnings: number;
  indeterminate: number;
};

export type FaultType = {
  Critical: number;
  Major: number;
  Minor: number;
  Warning: number;
  Indeterminate: number;
};

export type Faults = {
  faults: FaultsCountReturnType;
};